LayoutTests/webaudio/web-audio-is-playing-expected.txt

Testing Page::isPlayingAudio() with web audio

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.internals.pageMediaState().includes('IsPlayingAudio') is false
PASS window.internals.pageMediaState().includes('IsPlayingAudio') is false
PASS window.internals.pageMediaState().includes('IsPlayingAudio') became true
PASS successfullyParsed is true

TEST COMPLETE
